+++ src/main/kotlin/com/github/cplexopl/actions/OplRateAction.kt
@@ -1,0 +1,14 @@
+package com.github.cplexopl.actions
+
+import com.intellij.ide.BrowserUtil
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+
+class OplRateAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        BrowserUtil.browse("https://plugins.jetbrains.com/plugin/31125-cplex-opl")
+    }
+
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
+}