+++ src/test/kotlin/com/github/cplexopl/OplErrorReportSubmitterTest.kt
@@ -1,0 +1,15 @@
+package com.github.cplexopl
+
+import com.github.cplexopl.error.OplErrorReportSubmitter
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotNull
+
+class OplErrorReportSubmitterTest : BasePlatformTestCase() {
+
+    fun testErrorReportSubmitterActionText() {
+        val submitter = OplErrorReportSubmitter()
+        assertNotNull("ErrorReportSubmitter nie powinien być null", submitter)
+        assertEquals("Report Issue on GitHub", submitter.reportActionText)
+    }
+}